+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Version 2.0.0
 * ResultNodeInterface to ResultNode
 * CodeCoverageResultInterface to CodeCoverageResult
 * LineResultInterface to LineCountResult
+* CoverageResultInterface to CoverageResultNode
 
 Version 1.8.3
 ---------------------------------------------------------------